+++ client/src/pages/order/Order.jsx
@@ -1,5 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useState, useEffect } from "react";
+import { CardElement, useElements, useStripe } from '@stripe/react-stripe-js';
 import { Col, Row } from "antd";
 import equal from "fast-deep-equal";
 import { connect } from "react-redux";
@@ -19,11 +20,32 @@ import {
 import { BuyStep } from "../../components";
 import PaymentInput from '../../components/PaymentInput';
 import './styles.scss';
+import e from 'express';
 
 const style = {
   display: "block",
   lineHeight: "30px",
 };
+
+const options = {
+  hidePostalCode: true,
+  style: {
+    base: {
+      iconColor: "#000",
+      fontSize: "16px",
+      color: "#000",
+      letterSpacing: "0.025em",
+      "::placeholder": {
+        color: "#666"
+      }
+    },
+    invalid: {
+      iconColor: "red",
+      color: "red"
+    }
+  }
+}
+
 const Order = ({
   cartState,
   history,
@@ -32,6 +54,9 @@ const Order = ({
   orderProducts,
   orderProductsAuth,
 }) => {
+  const stripe = useStripe();
+  const elements = useElements();
+
   const [isProcessing, setIsProcessing] = useState(false);
   const [totalMoney, setTotalMoney] = useState(0);
   const [deliveryState, SetDeliveryState] = useState({
@@ -39,6 +64,17 @@ const Order = ({
     price: 35000,
   });
   const [paymentState, SetPaymentState] = useState(0);
+  const [isOpenStripe, setOpenStripe] = useState(false);
+  const [error, setError] = useState(null);
+  const [cardComplete, setCardComplete] = useState(false);
+  const [paymentMethod, setPaymentMethod] = useState(null);
+
+  const handleCardChange = (e) => {
+    console.log(e);
+    setError(e.error);
+    setCardComplete(e.complete);
+  }
+
   useEffect(() => {
     const cart = JSON.parse(localStorage.getItem("cart"));
     const guestInfo = JSON.parse(localStorage.getItem("guestInfo"));
@@ -78,6 +114,7 @@ const Order = ({
       });
     }
   }, [cartState, deliveryState]);
+  
   const onChangeDelivery = (e) => {
     SetDeliveryState({
       ...deliveryState,
@@ -85,6 +122,7 @@ const Order = ({
       price: e.target.value === 0 ? 35000 : 55000,
     });
   };
+
   const onChangePayment = (e) => {
     const paymentState = e.target.value;
     SetPaymentState(paymentState);
@@ -95,27 +133,54 @@ const Order = ({
       setOpenStripe(false);
     }
   };
-  const onFinish = async () => {
-    let cart = JSON.parse(localStorage.getItem("cart"));
-    if (!equal(cart, cartState)) {
-      return history.push("/cart");
-    }
-    setIsProcessing(true);
-    let res;
-    if (isAuthenticated) {
-      res = await orderProductsAuth(authState);
+  
+  const onFinish = async (e) => {
+    e.preventDefault();
+    const {error, paymentMethod} = await stripe.createPaymentMethod({
+      type: "card",
+      card: elements.getElement(CardElement)
+    });
+    if(!error) {
+      // try {
+      //     const {id} = paymentMethod
+      //     const response = await axios.post("http://localhost:4000/payment", {
+      //         amount: 1000,
+      //         id
+      //     })
+
+      //     if(response.data.success) {
+      //         console.log("Successful payment")
+      //         setSuccess(true)
+      //     }
+
+      // } catch (error) {
+      //     console.log("Error", error)
+      // }
+      console.log(paymentMethod);
     } else {
-      res = await orderProducts(guestState);
+        console.log(error.message)
     }
-    setIsProcessing(false);
-    if (res) {
-      store.dispatch({
-        type: REMOVE_CART,
-      });
-      store.dispatch({
-        type: CLEAR_CHECKOUT_INFO,
-      });
-    }
+    // let cart = JSON.parse(localStorage.getItem('cart'));
+    // if (!equal(cart, cartState)) {
+    //   return history.push('/cart');
+    // }
+    // setIsProcessing(true);
+    // let res;
+    // if (isAuthenticated) {
+    //   res = await orderProductsAuth(authState);
+    // } else {
+    //   res = await orderProducts(guestState);
+    // }
+    // setIsProcessing(false);
+    // if (res) {
+    //   store.dispatch({
+    //     type: REMOVE_CART,
+    //   });
+    //   store.dispatch({
+    //     type: CLEAR_CHECKOUT_INFO,
+    //   });
+    //   localStorage.removeItem('cart');
+    // }
   };
 
   return (
