+++ client/src/pages/order/styles.scss
@@ -61,3 +61,23 @@
         height: 50px;
     }
 }
+
+.payment-input {
+    padding: 10px;
+  }
+  
+  .StripeElement {
+    display: block;
+    margin: 10px 0 10px 0;
+    padding: 10px 14px;
+    font-size: 1em;
+    outline: 0;
+    box-shadow: 0 3px 5px -1px rgba(128, 147, 167, 0.2), 0 1px 18px 0 rgba(128, 147, 167, 0.12), 0 6px 10px 0 rgba(128, 147, 167, 0.14);
+    border-radius: 2px;
+  }
+  
+  .error-message {
+    text-align: center;
+    color: red;
+  }
+
